Update remaining yaml refs to toml crate (0.7.5)
