+++ src/bin/aardwolf-server.rs
@@ -1,11 +1,13 @@
-use std::error::Error;
+use std::error::Error as _;
 
 use aardwolf::{begin_log, configure};
-use clap::{load_yaml, App};
+use clap::{App};
 use config::Config;
+use toml::from_str;
 
-fn cli<'a, 'b>(yaml: &'a yaml_rust::yaml::Yaml) -> App<'a, 'b> {
-    App::from_yaml(yaml)
+fn cli(toml: &str) -> App {
+    let table: toml::value::Table = toml.parse().unwrap();
+    App::from(table)
         .name(env!("CARGO_PKG_NAME"))
         .version(env!("CARGO_PKG_VERSION"))
         .author(env!("CARGO_PKG_AUTHORS"))
@@ -33,8 +35,8 @@ mod actix {
 }
 
 fn main() -> Result<(), Box<dyn Error>> {
-    let yaml = load_yaml!("cli.yml");
-    let cli = cli(&yaml);
+    let toml = include_str!("cli.toml");
+    let cli = cli(toml);
     let config = configure(cli)?;
     let db_url = aardwolf::db_conn_string(&config)?;
